@@ -1,11 +1,9 @@
 /** @type {import('next').NextConfig} */
-
 const nextConfig = {
-    output: 'export',
-    basePath: '',
-    images: {
-        unoptimized: true
-    }
-}
+  output: "export",
+  images: {
+    unoptimized: true,
+  },
+};
 
-module.exports = nextConfig
+export default nextConfig;
