@@ -1,50 +1,27 @@
 "use client";
 
+import "@/i18n";
+import "../styles/index.css";
+import "../styles/prism-vsc-dark-plus.css";
+
+import { ThemeProvider } from "next-themes";
+
 import Footer from "@/components/Footer";
 import Header from "@/components/Header";
 import ScrollToTop from "@/components/ScrollToTop";
-import { ThemeProvider } from "next-themes";
-import "../styles/index.css";
-import "../styles/prism-vsc-dark-plus.css";
-import { useEffect, useState } from "react";
-import PreLoader from "@/components/Common/PreLoader";
 
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const [loading, setLoading] = useState<boolean>(true);
-
-  useEffect(() => {
-    setTimeout(() => setLoading(false), 1000);
-  }, []);
+export default function RootLayout({ children }: { children: React.ReactNode }) {
 
   return (
-    <html suppressHydrationWarning={true} className="!scroll-smooth" lang="en">
-      {/*
-        <head /> will contain the components returned by the nearest parent
-        head.js. Find out more at https://beta.nextjs.org/docs/api-reference/file-conventions/head
-      */}
+    <html suppressHydrationWarning className="!scroll-smooth" lang="en">
       <head />
-
       <body>
-        {loading ? (
-          <PreLoader />
-        ) : (
-          <>
-            <ThemeProvider
-              attribute="class"
-              enableSystem={false}
-              defaultTheme="light"
-            >
-              <Header />
-              {children}
-              <Footer />
-              <ScrollToTop />
-            </ThemeProvider>
-          </>
-        )}
+        <ThemeProvider attribute="class" enableSystem={false} defaultTheme="light">
+          <Header />
+          {children}
+          <Footer />
+          <ScrollToTop />
+        </ThemeProvider>
       </body>
     </html>
   );
