@@ -3,37 +3,37 @@ import { Menu } from "@/types/menu";
 const menuData: Menu[] = [
   {
     id: 1,
-    title: "Home",
+    title: "header.home",
     path: "/",
     newTab: false,
   },{
     id: 2,
-    title: "About",
+    title: "header.about",
     path: "/#about",
     newTab: false,
   },{
     id: 3,
-    title: "Media",
+    title: "header.media",
     path: "/media",
     newTab: false,
   },{
     id: 4,
-    title: "Contact",
+    title: "header.contact",
     path: "/#contact",
     newTab: false,
   },{
     id: 5,
-    title: "How To Register",
+    title: "header.howToRegister",
     path: "/register",
     newTab: false,
   },{
     id: 6,
-    title: "Login",
+    title: "header.login",
     path: "/app/auth/login",
     newTab: false,
   },{
     id: 7,
-    title: "Register",
+    title: "header.register",
     path: "/app/auth/register",
     newTab: false,
   },
