@@ -1,8 +1,13 @@
-import React from "react";
+"use client"
+
 import Image from "next/image";
 import Link from "next/link";
+import React from "react";
+import { useTranslation } from "react-i18next";
 
 const NotFound = () => {
+  const { t } = useTranslation();
+
   return (
     <section className="bg-white py-20 dark:bg-dark-2 lg:py-[110px]">
       <div className="container mx-auto">
@@ -57,17 +62,16 @@ const NotFound = () => {
                 </svg>
               </div>
               <h3 className="mb-5 text-2xl font-semibold text-dark dark:text-white">
-                We Can&#39;t Seem to Find The Page You&#39;re Looking For.
+                {t("notFound.title")}
               </h3>
               <p className="mb-8 text-base text-body-color dark:text-dark-6">
-                Oops! The page you are looking for does not exist. It might have
-                been moved or deleted.
+                {t("notFound.description")}
               </p>
               <Link
                 href="/"
                 className="rounded-md bg-dark px-7 py-3 text-base font-medium text-white transition hover:bg-primary dark:bg-primary dark:hover:bg-primary/80"
               >
-                Go To Home
+                {t("notFound.button")}
               </Link>
             </div>
           </div>
